fix: vm pool status on vmpools
